vupdated: handle signals even when disconnected
Currently, there's an issue where `vupdated` signal handling doesn't
work if it hasn't loaded an initial dial list from `VU-Server`.
Unfortunately, this describes the case where the system boots with no
dials connected, so if a laptop starts up without dials and then is
restarted, systemd will wait a minute and a half for `vupdated` to shut
itself down, making system shutdown slow. This fixes that by ensuring
that we always terminate immediately on `SIGTERM` or `SIGINT`, even if
we haven't got a dial list yet.